@@ -42,34 +42,34 @@
     "svelte": "^3.0.0 || ^4.0.0 || ^5.0.0"
   },
   "devDependencies": {
-    "@codemirror/lang-css": "^6.0.0",
-    "@codemirror/lang-html": "^6.0.0",
-    "@codemirror/lang-javascript": "^6.0.0",
-    "@codemirror/theme-one-dark": "^6.0.0",
-    "@eslint/compat": "^1.2.5",
-    "@eslint/js": "^9.18.0",
-    "@semantic-release/changelog": "^6.0.1",
-    "@semantic-release/commit-analyzer": "^11.1.0",
+    "@codemirror/lang-css": "^6.3.1",
+    "@codemirror/lang-html": "^6.4.9",
+    "@codemirror/lang-javascript": "^6.2.4",
+    "@codemirror/theme-one-dark": "^6.1.3",
+    "@eslint/compat": "^1.3.2",
+    "@eslint/js": "^9.35.0",
+    "@semantic-release/changelog": "^6.0.3",
+    "@semantic-release/commit-analyzer": "^13.0.1",
     "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^9.2.3",
-    "@semantic-release/npm": "^11.0.1",
+    "@semantic-release/github": "^11.0.5",
+    "@semantic-release/npm": "^12.0.2",
     "@sveltejs/adapter-static": "^3.0.9",
     "@sveltejs/kit": "^2.37.1",
     "@sveltejs/package": "^2.5.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
-    "codemirror": "^6.0.0",
+    "@sveltejs/vite-plugin-svelte": "^6.1.4",
+    "codemirror": "^6.0.2",
     "eslint": "^9.18.0",
-    "eslint-config-prettier": "^10.0.1",
-    "eslint-plugin-svelte": "^3.0.0",
-    "globals": "^16.0.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-svelte": "^3.12.2",
+    "globals": "^16.3.0",
     "prettier": "^3.6.2",
     "prettier-plugin-svelte": "^3.4.0",
-    "semantic-release": "^22.0.8",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.6.2",
-    "tslib": "^2.6.2",
-    "typescript": "^5.3.2",
-    "typescript-eslint": "^8.20.0",
-    "vite": "^5.0.0"
+    "semantic-release": "^24.2.7",
+    "svelte": "^5.38.7",
+    "svelte-check": "^4.3.1",
+    "tslib": "^2.8.1",
+    "typescript": "^5.9.2",
+    "typescript-eslint": "^8.43.0",
+    "vite": "^7.1.5"
   }
 }
